+++ include/ast/expressions.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "./expressions/expr.hpp"
+#include "./expressions/call.hpp"
+#include "./expressions/member_access.hpp"
+#include "./expressions/comparation.hpp"
+#include "./expressions/logical.hpp"
+
+
+#include "./expressions/import.hpp"
