--- include/ast/nodes/expressions.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include "ast/nodes/expressions/expr.hpp"
-#include "ast/nodes/expressions/call.hpp"
-#include "ast/nodes/expressions/member_access.hpp"
-#include "ast/nodes/expressions/comparation.hpp"
-#include "ast/nodes/expressions/logical.hpp"
-
-
-#include "ast/nodes/expressions/import.hpp"
