--- include/ast/nodes/statements.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "ast/nodes/statements/statement.hpp"
-#include "ast/nodes/statements/assignment.hpp"
-#include "ast/nodes/statements/loops.hpp"
-#include "ast/nodes/statements/conditional_branch.hpp"
