--- include/ast/nodes/statements/loops.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-#include "ast/nodes/statements/loops/loop.hpp"
-#include "ast/nodes/statements/loops/condition_loop.hpp"
