+++ include/ast/statements.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "./statements/statement.hpp"
+#include "./statements/assignment.hpp"
+#include "./statements/loops.hpp"
+#include "./statements/conditional_branch.hpp"
