+++ src/parser/expression/arith.cpp
@@ -1,5 +1,6 @@
 #include "parser/parse_arith.hpp"
 #include "ast/nodes.hpp"
+#include "ast/expressions/arith.hpp"
 #include "tokens/tokens.hpp"
 
 #include "parser/parse_prefix.hpp"
